Admin Scaffolding COMPLETE!
There's probably a lot that's broken. There is certainly a lot that
doesn't make sense (do we really need scaffolding to add news
comments?). Needless to say, this requires a schema update, and won't be
the last....
Also, found out that the reason why webroot wanted to update last commit
is because there was a recursive 'webroot' directory. DERP.